live_service/requirements_live_service.txt
duckdb>=1.1.0
numpy>=2.0.0
pandas>=2.2.0
PyYAML>=6.0.0
requests>=2.32.0

